requirements.txt
django==5.0.1
gunicorn==21.0.0
whitenoise[brotli]
djangorestframework==3.16.0